Get main todo list panel and add form working
Big commit: this adds the main App class, sets its constructor
to register all the Handlebars templates, get an initial todo
list from the server, and render it. It also handles display of
the add todo form modal, and submitting new todos via that form.
There's plenty of refactoring todo, and we may want to separate
the display logic into a display module.